Fix unit tests failing due to transitive database imports (#398)
Extract pure functions to separate modules to break transitive import
chains that were causing DATABASE_URL errors in unit tests:

- cleanEntryContent → src/server/feed/content-utils.ts
- parseFromAddress, stripEmailNameQuotes → src/server/email/parse-utils.ts
- findPermanentRedirectUrl, isHttpToHttpsUpgrade, REDIRECT_WAIT_PERIOD_MS
  → src/server/feed/redirect-utils.ts
- Worker core logic → src/server/jobs/worker-core.ts

Original modules re-export for backwards compatibility. Tests now import
directly from the new pure modules that don't depend on the database.
